Artifacts singleton moved to make quicktest work again
