+++ packages/protocol/lib/artifactsSingleton.ts
