+++ packages/protocol/lib/test-utils.ts
@@ -2,7 +2,7 @@
 import { hasEntryInRegistry, usesRegistry } from '@celo/protocol/lib/registry-utils'
 import { getParsedSignatureOfAddress } from '@celo/protocol/lib/signing-utils'
 import { getDeployedProxiedContract } from '@celo/protocol/lib/web3-utils'
-import { ArtifactsSingleton } from '@celo/protocol/migrations/artifactsSingleton'
+import { ArtifactsSingleton } from '@celo/protocol/lib/artifactsSingleton'
 import { config } from '@celo/protocol/migrationsConfig'
 import { privateKeyToAddress } from '@celo/utils/lib/address'
 import { soliditySha3 } from '@celo/utils/lib/solidity'
