+++ packages/protocol/lib/web3-utils.ts
@@ -15,7 +15,7 @@ import { GoldTokenInstance, MultiSigInstance, OwnableInstance, ProxyContract, Pr
 import { StableTokenInstance } from 'types/mento'
 import Web3 from 'web3'
 import { ContractPackage } from '../contractPackages'
-import { ArtifactsSingleton } from '../migrations/artifactsSingleton'
+import { ArtifactsSingleton } from './artifactsSingleton'
 
 const truffleContract = require('@truffle/contract');
