+++ packages/protocol/test/common/migration.ts
@@ -5,7 +5,7 @@ import {
 } from '@celo/protocol/lib/test-utils'
 import { getDeployedProxiedContract } from '@celo/protocol/lib/web3-utils'
 import { ContractPackage } from 'contractPackages'
-import { ArtifactsSingleton } from '../../migrations/artifactsSingleton'
+import { ArtifactsSingleton } from '../../lib/artifactsSingleton'
 
 const getProxiedContract = async (contractName: string, contractPackage: ContractPackage) => {
   const artifactsObject = ArtifactsSingleton.getInstance(contractPackage, artifacts)
